Extract Skills to own component
