+++ app/_components/IncomingData/TextBlock.module.css
@@ -1,0 +1,3 @@
+.container:not(:last-child) {
+  margin-bottom: 1.25rem 
+}
